Introduce rspec helpers
Creates template and subject methods for rspec.

The template uses a mock unless ActionView is available. It's
recommended to require 'action_view' only if you want to tests methods
that depends on the template (like number_to_currency for example)

The subject method expects that you define a `object` let which will
reflect the attributes the class being presented. It's recommended to
use a stub for test in isolation